Split MCPServer.swift into per-tool handler files behind ToolHandler
Step #1 of the architectural roadmap. The previous shape had three
sources of truth that had to stay in lockstep: the `ToolName` enum,
the `mcpToolSchemas()` array in MCPToolSchemas.swift, and the
`switch tool` dispatch in `configureMCPServer`. Adding a tool meant
touching three places.

Introduce a `ToolHandler` protocol with `static var name`, `static var
schema`, and `static func handle(_:ctx:)`. Each of the 12 tools moves
to its own file under `Sources/PreviewsCLI/Handlers/`, carrying the
schema next to the implementation. `configureMCPServer` derives both
the `ListTools` response and the `CallTool` dispatch from a single
`handlerRegistry` array — the array IS the inventory.

`HandlerContext` bundles the engine-layer dependencies (host, iosState,
configCache, router, macCompiler, server) so handlers receive deps via
parameter rather than reaching for file-private vars. Static handler
types are `Sendable`-trivially since they have no instance state.

`MCPServer.swift` shrinks from 1131 to 92 lines (target was <200);
shared infrastructure moves to `MCPServerSupport.swift` (heartbeat,
progress reporter, version advertisement, `parseTraits`,
`configQualityForSession`); `MCPToolSchemas.swift` is deleted.

The acceptance gate from the architectural plan — "ListTools response
bytes are identical pre/post" — is enforced by a new
`ListToolsSnapshotTests` snapshot that pins the JSON-encoded schemas
against a fixture file. The pre-refactor bytes were captured on the
prior commit, and the test passes against the post-refactor registry.
